requirement.txt
numpy
pandas
matplotlib
python-dotenv
mysql-connector-python
pymysql
scikit-learn
